make indirect functions lazy
